refactor: Checks that errors in `catch` claues are of `Error` type before accesing `Error` specific properties and preferes `delete` to setting optional values to `undefined`
With typescript 4.4 strict mode, catch clauses error argument is now `unknown` by default which means we need to add type guards to be the most typesafe.

It also brings `exactOptionalPropertyTypes` which prevents setting optional properties to undefined when we really just want them to not be required to be set. We can't set the flag yet but we changed breakages in order to update later.