fix(rikaicontent): Properly check for text areas and input elements when deciding to save cursor information on mousedown (#1282)
The previous logic required the text box to be in a `<form>` element which is not often true.

A follow up to this should be to also handle other cases like `content: editable`. I'm not sure how that works now though so maybe it's fine.

Also, update the timeout for data_test.ts since it wasn't long enough to complete on the normal Github Codespace.