refactor(makeText): Rename variables, move variables to local scope (#1623)
The goal of this pull request is to give variables in `makeText` more descriptive names so that the code is easier to understand. Some variables were also put into a more local scope so that they're easier to follow.

This pull request does not change the functionality of `makeText` - the copy-to-clipboard feature works exactly how it did before this pull request.

I tried to make this pull request have a small code diff so that it's easier to review. I plan on making more changes to `makeText` in subsequent pull requests which will make it even easier to understand.

Changes I plan on making in subsequent pull requests:
- Restructure the `kanjiInfoLabelList` array so that it's an array of objects - so its type might change to something like `{ kanjiInfoCode: string; kanjiInfoName: string }[]` so that it's easier to understand.
- Write a dict entry parsing function which is well-tested and unifies the parsing logic used in `makeText`, `makeHtml`, and the update-db.ts file.
- Change the `DictEntryData` type to represent the valid states of the application better - something like `type DictEntryData = KanjiDictEntry | WordDictEntry;`