fix(google-docs): Set docs_annotate_canvas property in a MV3 compliant manner
New CSP changes means we can no longer inject a script into the MAIN world from the ISOLATED world.

Instead, we use the new `world` key to inject a simple script directly into the MAIN world.

One caveat of this is that we can no longer read the extension ID dynamically (as it's not available in MAIN world) so instead we hardcode it into the script as it extremely stable.

Fixes #2175