fix: restore the accurate length counting functionality
The past code had trueLen to keep track of the actual length of the input string at each iteration of the
word conversion loop. This could happen due to special characters that we skip as well as the fact that half width
katakana treats voiced/semi-voiced characters as two characters instead of one.

Will take some thought to figure out exactly how to do it cleanly.